refactor(contract): add api module root and lift shared test fixtures
